Fix: single quote index 싱글 쿼트 인덱스 -1
